# chore: faster usePerpsMarket hook